Extracted interval's logic to external file
