+++ src/components/Header/index.tsx
@@ -2,6 +2,8 @@ import React, { useContext, useEffect, useState } from 'react';
 import Swal from 'sweetalert2'
 
 import * as S from './styled';
+import { handleInterval } from './interval';
+
 import UserContext, { CountdownValueContext } from '../../state/context';
 
 const MM_SS_REGEX = /^(?:(?:([01]?\d|2[0-3]):)?([0-5]?\d))$/;
@@ -21,27 +23,10 @@ function Header() {
   useEffect(() => {
     if (!started) return;
 
-    const interval = setInterval(() => {
-      const [ minutes, seconds ] = countdownValue.split(':');
-
-      let minInt = parseInt(minutes);
-      let secInt = parseInt(seconds);
-
-      if (secInt - 1 >= 0) {
-        secInt -= 1;
-      } else {
-        secInt = 59;
-        minInt = Math.max(minInt - 1, 0);
-      }
-
-      const newTime = `${minInt}`.padStart(2, '0')
-        + ':' + `${secInt}`.padStart(2, '0');
-      setCurrentValue(newTime);
-
-      if (minInt === 0 && secInt === 0) {
-        console.log('Times up');
-      }
-    }, 1000);
+    const interval = setInterval(
+      () => handleInterval(countdownValue, setCurrentValue),
+      1000
+    );
 
     return () => clearInterval(interval);
   }, [countdownValue, setCurrentValue, started]);
